Change "share of own..." to "fraction own..."
This is to align the code with the current drafts figures. IMHO this still sub-optimal, as it leaves the unfortunate ambiguity whether "O" mean own or other's in the short form FOC and it resuts in  too long axis labels. I would prefer "Own choices [%]" and a scale from 0-100 instead of 0-1. And switching from Own/Other to maybe Preferred/Non-Preferred P/N (should not use R or L to keep the side choices unambiguous).