Use knowledge about the directory structure inside the SESSIONLOGS directory to find relevant files/folders.
The ols find_all_files does as it claims, but its run time sclaes badly with the size of the searched directory tree. Instead use the knowledge about how SCP_01 saves data and only search the relevant directories. Also return a list of the sissiondir FQNs and a logical vector which contain valid SCP tankdirectories and a list of session IDs.